Implement area address TLV
