+++ protocols/isis/server/server.go
@@ -6,7 +6,6 @@ import (
 	"time"
 
 	"github.com/bio-routing/bio-rd/protocols/device"
-	"github.com/bio-routing/bio-rd/protocols/isis/packet"
 	"github.com/bio-routing/bio-rd/protocols/isis/types"
 	btime "github.com/bio-routing/bio-rd/util/time"
 )
@@ -103,27 +102,4 @@ func netsCompatible(nets []*types.NET) bool {
 	}
 
 	return true
-}
-
-func (s *Server) generateL2LSP(sequenceNumber uint32) {
-	lsp := &packet.LSPDU{
-		RemainingLifetime: 3600,
-		LSPID: packet.LSPID{
-			SystemID:     s.nets[0].SystemID,
-			PseudonodeID: 0,
-			LSPNumber:    0,
-		},
-		SequenceNumber: sequenceNumber,
-	}
-
-	lsp.TypeBlock |= 0x3 // level2, last two bits
-	lsp.UpdateLength()
-	lsp.SetChecksum()
-
-	s.lsdbL2.lspsMu.Lock()
-	defer s.lsdbL2.lspsMu.Unlock()
-
-	s.lsdbL2.lsps[lsp.LSPID] = newLSDBEntry(lsp)
-
-	// TODO: Set SRM?
 }
